Split up NotificationScreen.tsx
