+++ src/navigation/root/NotificationScreen/NotificationRow/index.ts
@@ -1,0 +1,1 @@
+export { AnimatedNotificationRow as NotificationRow } from "./NotificationRow";
